introduce more iterators
There may still be some destabilization.  Major introduction is iterators in the xml level.

+++ svg/svgattributes.h
@@ -407,7 +407,7 @@ namespace waavs {
         
         void drawSelf(IRenderSVG* ctx, IAmGroot* groot) override
         {
-            ctx->fontSize(fValue);
+            ctx->fontSize(static_cast<float>(fValue));
         }
     };
